requirements.txt
# Axiom / NearTax Python Dependencies

# Database
alembic>=1.13.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

# HTTP clients
aiohttp>=3.9.0
requests>=2.31.0

# Environment
python-dotenv>=1.0.0

# FastAPI web framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# Auth
webauthn>=2.0.0
itsdangerous>=2.1.0

# Cloud / storage
boto3>=1.34.0

# FastAPI testing
httpx>=0.27.0
aiofiles>=24.0.0

# Rate limiting
slowapi>=0.1.9
